+++ C & C++/BilgisayarMühendisligineGiris/Week 10/TekCift.c
