+++ C & C++/BilgisayarMühendisligineGiris/Week 3/ödev-2.c
