+++ C & C++/BilgisayarMühendisligineGiris/Week 5/ödev 2.c
