+++ C & C++/BilgisayarMühendisligineGiris/Week 7/tek_çift_sayilari.c
