+++ C & C++/Kodumun Yeri/BasitResim.c
