+++ C & C++/Kodumun Yeri/Gülücük.c
